requirements.txt
flask
python-dotenv
supabase
mcp-server